billing: refactoring domain
