+++ internal/services/billing/application/account/account.go
@@ -6,7 +6,7 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/batazor/shortlink/internal/pkg/logger"
-	billing "github.com/batazor/shortlink/internal/services/billing/domain/billing/v1"
+	billing "github.com/batazor/shortlink/internal/services/billing/domain/billing/account/v1"
 	billing_store "github.com/batazor/shortlink/internal/services/billing/infrastructure/store"
 )
