+++ internal/services/billing/domain/billing/balance/v1/balance.proto
@@ -1,10 +1,16 @@
 syntax = "proto3";
 
-package domain.billing.v1;
+package domain.billing.balance.v1;
 
 option go_package = "github.com/batazor/shortlink/internal/services/billing/domain/billing/v1";
 
 message Balance {
   string payment_id = 1;
   int64 amount = 2;
+}
+
+enum Event {
+  EVENT_UNSPECIFIED = 0;
+
+  EVENT_BALANCE_UPDATE = 1;
 }
