--- internal/services/billing/domain/billing/v1/event.proto
@@ -1,35 +1,0 @@
-syntax = "proto3";
-
-package domain.billing.v1;
-
-option go_package = "github.com/batazor/shortlink/internal/services/billing/domain/billing/v1";
-
-enum Event {
-  EVENT_UNSPECIFIED = 0;
-
-  // Payment
-  EVENT_PAYMENT_NEW = 1;
-  EVENT_PAYMENT_PENDING = 2;
-  EVENT_PAYMENT_APPROVE = 3;
-  EVENT_PAYMENT_CLOSE = 4;
-
-  // Balance
-  EVENT_BALANCE_UPDATE = 10;
-
-  // Order
-  EVENT_ORDER_NEW = 20;
-  EVENT_ORDER_UPDATE = 21;
-  EVENT_ORDER_APPROVED = 22;
-  EVENT_ORDER_PAID = 23;
-  EVENT_ORDER_CLOSE = 24;
-  EVENT_ORDER_REJECT = 25;
-
-  // Account
-  EVENT_ACCOUNT_NEW = 40;
-  EVENT_ACCOUNT_DELETE = 41;
-
-  // Tariff
-  EVENT_TARIFF_NEW = 50;
-  EVENT_TARIFF_UPDATE = 51;
-  EVENT_TARIFF_CLOSE = 52;
-}
