+++ internal/services/billing/infrastructure/api/rpc/balance/v1/balance_rpc.proto
@@ -5,8 +5,8 @@ package infrastructure.api.rpc.balance.v1;
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/balance_rpc";
 
 import "google/protobuf/empty.proto";
-import "domain/billing/v1/balance.proto";
-import "domain/billing/v1/payment.proto";
+import "domain/billing/balance/v1/balance.proto";
+import "domain/billing/payment/v1/payment.proto";
 
 service BalanceService {
   rpc BalanceHistory(google.protobuf.Empty) returns(BalanceHistoryResponse) {}
@@ -14,13 +14,13 @@ service BalanceService {
 }
 
 message BalanceUpdateRequest {
-  domain.billing.v1.Balance balance = 1;
+  domain.billing.balance.v1.Balance balance = 1;
 }
 
 message BalanceUpdateResponse {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
 
 message BalanceHistoryResponse {
-  repeated domain.billing.v1.Balance list = 1;
+  repeated domain.billing.balance.v1.Balance list = 1;
 }
