+++ internal/services/billing/infrastructure/api/rpc/order/v1/order_rpc.proto
@@ -5,7 +5,7 @@ package infrastructure.api.rpc.order.v1;
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/order_rpc";
 
 import "google/protobuf/empty.proto";
-import "domain/billing/v1/order.proto";
+import "domain/billing/order/v1/order.proto";
 
 service OrderService {
   rpc OrderHistory(google.protobuf.Empty) returns(OrderHistoryResponse) {}
@@ -16,38 +16,38 @@ service OrderService {
 }
 
 message OrderHistoryResponse {
-  repeated domain.billing.v1.Order list = 1;
+  repeated domain.billing.order.v1.Order list = 1;
 }
 
 message OrderCreateRequest {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderCreateResponse {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderUpdateRequest {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderUpdateResponse {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderCloseRequest {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderCloseResponse {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderApproveRequest {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
 
 message OrderApproveResponse {
-  domain.billing.v1.Order order = 1;
+  domain.billing.order.v1.Order order = 1;
 }
