+++ internal/services/billing/infrastructure/api/rpc/payment/v1/billing_rpc.proto
@@ -5,7 +5,7 @@ package infrastructure.api.rpc.payment.v1;
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/payment_rpc";
 
 import "google/protobuf/empty.proto";
-import "domain/billing/v1/payment.proto";
+import "domain/billing/payment/v1/payment.proto";
 
 service PaymentService {
   rpc Payment(PaymentRequest) returns(PaymentResponse) {}
@@ -15,29 +15,29 @@ service PaymentService {
 }
 
 message PaymentRequest {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
 
 message PaymentResponse {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
 
 message PaymentsResponse {
-  repeated domain.billing.v1.Payment list = 1;
+  repeated domain.billing.payment.v1.Payment list = 1;
 }
 
 message PaymentCreateRequest {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
 
 message PaymentCreateResponse {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
 
 message PaymentCloseRequest {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
 
 message PaymentCloseResponse {
-  domain.billing.v1.Payment payment = 1;
+  domain.billing.payment.v1.Payment payment = 1;
 }
