+++ internal/services/billing/infrastructure/api/rpc/tariff/v1/tariff_rpc.proto
@@ -5,7 +5,7 @@ package infrastructure.api.rpc.tariff.v1;
 option go_package = "github.com/batazor/shortlink/internal/billing/infrastructure/tariff_rpc";
 
 import "google/protobuf/empty.proto";
-import "domain/billing/v1/tariff.proto";
+import "domain/billing/tariff/v1/tariff.proto";
 
 service TariffService {
   rpc Tariff(TariffRequest) returns(TariffResponse) {}
@@ -16,37 +16,37 @@ service TariffService {
 }
 
 message TariffRequest {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffResponse {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffsResponse {
-  repeated domain.billing.v1.Tariff list = 1;
+  repeated domain.billing.tariff.v1.Tariff list = 1;
 }
 
 message TariffCreateRequest {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffCreateResponse {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffUpdateRequest {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffUpdateResponse {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffCloseRequest {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
 
 message TariffCloseResponse {
-  domain.billing.v1.Tariff tariff = 1;
+  domain.billing.tariff.v1.Tariff tariff = 1;
 }
