+++ internal/services/billing/infrastructure/store/postgres/account.go
@@ -9,7 +9,7 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 
 	"github.com/batazor/shortlink/internal/pkg/db"
-	v1 "github.com/batazor/shortlink/internal/services/billing/domain/billing/v1"
+	"github.com/batazor/shortlink/internal/services/billing/domain/billing/account/v1"
 )
 
 var (
