+++ internal/services/billing/infrastructure/store/postgres/order.go
@@ -6,7 +6,7 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 
 	"github.com/batazor/shortlink/internal/pkg/db"
-	v1 "github.com/batazor/shortlink/internal/services/billing/domain/billing/v1"
+	"github.com/batazor/shortlink/internal/services/billing/domain/billing/order/v1"
 )
 
 type Order struct {
