+++ internal/services/billing/infrastructure/store/postgres/payment.go
@@ -6,7 +6,7 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 
 	"github.com/batazor/shortlink/internal/pkg/db"
-	v1 "github.com/batazor/shortlink/internal/services/billing/domain/billing/v1"
+	"github.com/batazor/shortlink/internal/services/billing/domain/billing/payment/v1"
 )
 
 type Payment struct {
